Extracted mapping and scoring logic into separete services
