+++ parkulator-backend/src/main/java/hr/parkulator/parkulator_backend/controller/ParkingController.java
@@ -3,9 +3,9 @@ package hr.parkulator.parkulator_backend.controller;
 import hr.parkulator.parkulator_backend.dto.parking.ParkingDTO;
 import hr.parkulator.parkulator_backend.dto.parking.ParkingSearchResponseDTO;
 import hr.parkulator.parkulator_backend.entities.Parking;
-import hr.parkulator.parkulator_backend.services.ParkingService;
-
 import hr.parkulator.parkulator_backend.services.DataServices.ParkingDataService;
+import hr.parkulator.parkulator_backend.services.ParkingServices.ParkingService;
+
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
